+++ Magic/src/main/java/com/elmakers/mine/bukkit/integration/SkillAPIManager.java
@@ -1,15 +1,18 @@
 package com.elmakers.mine.bukkit.integration;
 
+import com.elmakers.mine.bukkit.magic.ManaController;
 import com.sucy.skill.SkillAPI;
+import com.sucy.skill.api.player.PlayerClass;
 import com.sucy.skill.api.player.PlayerData;
 import com.sucy.skill.manager.AttributeManager;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 
+import java.util.Collection;
 import java.util.Map;
 import java.util.Set;
 
-public class SkillAPIManager {
+public class SkillAPIManager implements ManaController {
     private final Plugin skillAPIPlugin;
     private final Plugin owningPlugin;
 
@@ -44,5 +47,45 @@ public class SkillAPIManager {
         if (attributeManager == null) return null;
         PlayerData playerData = SkillAPI.getPlayerData(player);
         return playerData.getAttributes();
+    }
+
+    @Override
+    public int getMaxMana(Player player) {
+        PlayerData playerData = SkillAPI.getPlayerData(player);
+        return (int)playerData.getMaxMana();
+    }
+
+    @Override
+    public int getManaRegen(Player player) {
+        PlayerData playerData = SkillAPI.getPlayerData(player);
+        Collection<PlayerClass> classes = playerData.getClasses();
+        double amount = 0;
+        for (PlayerClass c : classes)
+        {
+            if (c.getData().hasManaRegen())
+            {
+                amount += c.getData().getManaRegen();
+            }
+        }
+
+        return (int)amount;
+    }
+
+    @Override
+    public float getMana(Player player) {
+        PlayerData playerData = SkillAPI.getPlayerData(player);
+        return (float)playerData.getMana();
+    }
+
+    @Override
+    public void removeMana(Player player, float amount) {
+        PlayerData playerData = SkillAPI.getPlayerData(player);
+        playerData.setMana(Math.max(0, playerData.getMana() - amount));
+    }
+
+    @Override
+    public void setMana(Player player, float amount) {
+        PlayerData playerData = SkillAPI.getPlayerData(player);
+        playerData.setMana(amount);
     }
 }